Address review comments. Write a few better comments inside tests.
